move skill components into its own folder
